[241014] 4. 게시글목록[최종조립]     - 게시글목록페이지 src/app/boards/page.tsx 경로의 파일을 수정합니다.         - 당 페이지에 위에서 만든 2개의 컴포넌트(게시글목록, 페이지내이션)를 불러와서 조립합니다.         - state를 게시글목록페이지로 끌어올려 data, refetch, lastPage 등의 변수 및 함수를 두 컴포넌트가 적절히 공유할 수 있도록 합니다. 5. 컴포넌트[리팩토링]     - 페이지내이션 컴포넌트의 파일을 보완해 주세요.         - 타입에러가 감지되어 빨간 밑줄이 그어지는 부분에 타입스크립트를 적용하여 문제를 해결해 주세요.         - 유지보수가 쉽도록 파일을 hook.ts, index.tsx, queries.ts, styles.ts, types.ts 로 분리해 주세요.
