+++ 02-main_camp/Day31/src/app/boards/styles.module.css
@@ -1,8 +1,15 @@
 .pageContainer {
-  width: 100%;
+  max-width: 80rem;
+
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin-top: 2.5rem;
-  margin-bottom: 2.5rem;
+  gap: 1.5rem;
+
+  padding: 1.5rem 3rem 1.5rem 3rem;
+
+  border-radius: 16px;
+
+  background-color: #ffffff;
+  box-shadow: 0px 0px 20px 0px #00000014;
 }
