+++ 02-main_camp/Day31/src/app/components/boards/Boards-list/list/components/Pagination/types.ts
@@ -1,0 +1,14 @@
+import { Exact, FetchBoardsQuery, InputMaybe } from "@/commons/gql/graphql";
+import { ApolloQueryResult } from "@apollo/client";
+
+export interface IPaginationProps {
+  refetch: (
+    variables?:
+      | Partial<
+          Exact<{
+            page?: InputMaybe<number> | undefined;
+          }>
+        >
+      | undefined
+  ) => Promise<ApolloQueryResult<FetchBoardsQuery>>;
+}
