+++ 02-main_camp/Day31/src/app/components/boards/Boards-list/list/hook.tsx
@@ -1,10 +1,6 @@
 import { FetchBoardsDocument } from "@/commons/gql/graphql";
 import { useQuery } from "@apollo/client";
 
-const useBoardsList = () => {
-  const { data } = useQuery(FetchBoardsDocument);
-
-  return { data };
-};
+const useBoardsList = () => {};
 
 export default useBoardsList;
