+++ 02-main_camp/Day31/src/app/components/boards/Boards-list/list/types.ts
@@ -1,0 +1,5 @@
+import { FetchBoardsQuery } from "@/commons/gql/graphql";
+
+export interface IBoardsListProps {
+  data?: FetchBoardsQuery;
+}
