+++ 02-main_camp/Day31/src/app/globals.css
@@ -11,8 +11,6 @@
 
 body {
   width: 100vw;
-  padding-bottom: 2.5rem;
-
   display: flex;
   flex-direction: column;
   align-items: center;
