+++ 02-main_camp/Day31/src/commons/layout/banner/styles.module.css
@@ -1,3 +1,4 @@
 .BannerContainer {
   width: 100%;
+  height: 32rem;
 }
