+++ 02-main_camp/Day31/src/commons/layout/styles.module.css
@@ -4,3 +4,12 @@
   flex-direction: column;
   align-items: center;
 }
+
+.childrenContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 2.5rem;
+  margin-bottom: 2.5rem;
+}
